third_party/blink/web_tests/platform/mac/external/wpt/editing/run/strikethrough_1001-2000-expected.txt

This is a testharness.js-based test.
Found 171 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<strike>foo[bar]baz</strike>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<strike>foo[bar]baz</strike>" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<strike>foo[bar]baz</strike>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<strike>foo[b<span style=\\"color:blue\\">ar]ba</span>z</strike>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"text-decoration:line-through\\">foo</span>b<span style=\\"color:rgb(0, 0, 255)\\">ar<span style=\\"text-decoration:line-through\\">ba</span></span><span style=\\"text-decoration:line-through\\">z</span>" but got "<strike>foo</strike>b<span style=\\"color:rgb(0, 0, 255)\\">ar<span style=\\"text-decoration-line:line-through\\">ba</span></span><span style=\\"text-decoration-line:line-through\\">z</span>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<strike>foo[b<span style=\\"color:blue\\">ar]ba</span>z</strike>" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<strike>foo[b<span style=\\"color:blue\\">ar]ba</span>z</strike>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<strike>foo[b<span style=\\"color:blue\\">ar]ba</span>z</strike>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<strike>foo[b<span style=\\"color:blue\\">ar]ba</span>z</strike>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<strike>foo[b<span style=\\"color:blue\\">ar]ba</span>z</strike>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<strike>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</strike>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"text-decoration:line-through\\">foo</span>b<span style=\\"color:rgb(0, 0, 255)\\" id=\\"foo\\">ar<span style=\\"text-decoration:line-through\\">ba</span></span><span style=\\"text-decoration:line-through\\">z</span>" but got "<strike>foo</strike>b<span style=\\"color:rgb(0, 0, 255)\\" id=\\"foo\\">ar<span style=\\"text-decoration-line:line-through\\">ba</span></span><span style=\\"text-decoration-line:line-through\\">z</span>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<strike>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</strike>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<strike>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</strike>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<strike>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</strike>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<strike>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</strike>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<strike>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</strike>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"text-decoration:line-through\\">foo</span>b<span style=\\"font-size:3em\\">ar<span style=\\"text-decoration:line-through\\">ba</span></span><span style=\\"text-decoration:line-through\\">z</span>" but got "<strike>foo</strike>b<span style=\\"font-size:3em\\">ar<span style=\\"text-decoration-line:line-through\\">ba</span></span><span style=\\"text-decoration-line:line-through\\">z</span>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<strike>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</strike>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<strike>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</strike>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<strike>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</strike>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<strike>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</strike>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<strike>foo[b<i>ar]ba</i>z</strike>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"text-decoration:line-through\\">foo</span>b<i>ar<span style=\\"text-decoration:line-through\\">ba</span></i><span style=\\"text-decoration:line-through\\">z</span>" but got "<strike>foo</strike>b<i style=\\"\\">ar<span style=\\"text-decoration-line:line-through\\">ba</span></i><span style=\\"text-decoration-line:line-through\\">z</span>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<strike>foo[b<i>ar]ba</i>z</strike>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<strike>foo[b<i>ar]ba</i>z</strike>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<strike>foo[b<i>ar]ba</i>z</strike>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<strike>foo</strike>b<i>ar<strike>ba</strike></i><strike>z</strike>" but got "<strike>foo</strike>b<i style=\\"\\">ar<strike>ba</strike></i><strike>z</strike>"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<strike>foo[b<i>ar]ba</i>z</strike>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<strike>foo[b<i>ar]ba</i>z</strike>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<ins>[bar]</ins>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<ins style=\\"text-decoration:underline line-through\\">bar</ins>baz" but got "foo<ins style=\\"text-decoration-line:line-through\\">bar</ins>baz"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<ins>[bar]</ins>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<ins>[bar]</ins>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<ins>[bar]</ins>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<ins>[bar]</ins>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<ins>foo[bar]baz</ins>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ins>foo<span style=\\"text-decoration:line-through\\">bar</span>baz</ins>" but got "<ins>foo<span style=\\"text-decoration-line:line-through\\">bar</span>baz</ins>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<ins>foo[bar]baz</ins>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<ins>foo[bar]baz</ins>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<ins>foo[bar]baz</ins>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<ins>foo[bar]baz</ins>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<ins>foo[b<span style=\\"color:blue\\">ar]ba</span>z</ins>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ins>foo<span style=\\"text-decoration:line-through\\">b</span><span style=\\"color:rgb(0, 0, 255)\\"><span style=\\"text-decoration:line-through\\">ar</span>ba</span>z</ins>" but got "<ins>foo<span style=\\"text-decoration-line:line-through\\">b</span><span style=\\"color:rgb(0, 0, 255)\\"><span style=\\"text-decoration-line:line-through\\">ar</span>ba</span>z</ins>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<ins>foo[b<span style=\\"color:blue\\">ar]ba</span>z</ins>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<ins>foo[b<span style=\\"color:blue\\">ar]ba</span>z</ins>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<ins>foo[b<span style=\\"color:blue\\">ar]ba</span>z</ins>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<ins>foo[b<span style=\\"color:blue\\">ar]ba</span>z</ins>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<ins>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</ins>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ins>foo<span style=\\"text-decoration:line-through\\">b</span><span style=\\"color:rgb(0, 0, 255)\\" id=\\"foo\\"><span style=\\"text-decoration:line-through\\">ar</span>ba</span>z</ins>" but got "<ins>foo<span style=\\"text-decoration-line:line-through\\">b</span><span style=\\"color:rgb(0, 0, 255)\\" id=\\"foo\\"><span style=\\"text-decoration-line:line-through\\">ar</span>ba</span>z</ins>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<ins>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</ins>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<ins>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</ins>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<ins>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</ins>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<ins>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</ins>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<ins>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</ins>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ins>foo<span style=\\"text-decoration:line-through\\">b</span><span style=\\"font-size:3em\\"><span style=\\"text-decoration:line-through\\">ar</span>ba</span>z</ins>" but got "<ins>foo<span style=\\"text-decoration-line:line-through\\">b</span><span style=\\"font-size:3em\\"><span style=\\"text-decoration-line:line-through\\">ar</span>ba</span>z</ins>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<ins>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</ins>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<ins>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</ins>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<ins>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</ins>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<ins>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</ins>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<ins>foo[b<i>ar]ba</i>z</ins>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ins>foo<span style=\\"text-decoration:line-through\\">b</span><i><span style=\\"text-decoration:line-through\\">ar</span>ba</i>z</ins>" but got "<ins>foo<span style=\\"text-decoration-line:line-through\\">b</span><i><span style=\\"text-decoration-line:line-through\\">ar</span>ba</i>z</ins>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<ins>foo[b<i>ar]ba</i>z</ins>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<ins>foo[b<i>ar]ba</i>z</ins>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<ins>foo[b<i>ar]ba</i>z</ins>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<ins>foo[b<i>ar]ba</i>z</ins>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "foo<del>[bar]</del>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "foo<del>[bar]</del>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "<del>foo[bar]baz</del>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "<del>foo[bar]baz</del>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "<del>foo[b<span style=\\"color:blue\\">ar]ba</span>z</del>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "<del>foo[b<span style=\\"color:blue\\">ar]ba</span>z</del>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "<del>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</del>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "<del>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</del>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "<del>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</del>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "<del>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</del>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "<del>foo[b<i>ar]ba</i>z</del>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "<del>foo[b<i>ar]ba</i>z</del>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "foo<span style=\\"text-decoration: underline line-through\\">[bar]</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"text-decoration:underline\\">bar</span>baz" but got "foo<u>bar</u>baz"
[FAIL] [["strikethrough",""]] "foo<span style=\\"text-decoration: underline line-through\\">[bar]</span>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "foo<span style=\\"text-decoration: underline line-through\\">[bar]</span>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<span style=\\"text-decoration: underline line-through\\">b[a]r</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"text-decoration:underline\\"><span style=\\"text-decoration:line-through\\">b</span>a<span style=\\"text-decoration:line-through\\">r</span></span>baz" but got "foo<span style=\\"text-decoration:underline line-through\\">b</span><span style=\\"text-decoration-line:underline\\">a</span><span style=\\"text-decoration:underline line-through\\">r</span>baz"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<span style=\\"text-decoration: underline line-through\\">b[a]r</span>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<span style=\\"text-decoration: underline line-through\\">b[a]r</span>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<span style=\\"text-decoration: underline line-through\\">b[a]r</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"text-decoration:underline\\"><strike>b</strike>a<strike>r</strike></span>baz" but got "foo<span style=\\"text-decoration:underline line-through\\">b</span><u>a</u><span style=\\"text-decoration:underline line-through\\">r</span>baz"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<span style=\\"text-decoration: underline line-through\\">b[a]r</span>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<span style=\\"text-decoration: underline line-through\\">b[a]r</span>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<s style=\\"text-decoration: underline\\">[bar]</s>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<s style=\\"text-decoration: underline\\">[bar]</s>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<s style=\\"text-decoration: underline\\">[bar]</s>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<s style=\\"text-decoration: underline\\">[bar]</s>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<s style=\\"text-decoration: underline\\">b[a]r</s>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<s style=\\"text-decoration: underline\\">b[a]r</s>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<s style=\\"text-decoration: underline\\">b[a]r</s>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<s style=\\"text-decoration: underline\\">b[a]r</s>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<u style=\\"text-decoration: line-through\\">[bar]</u>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<u style=\\"text-decoration: line-through\\">[bar]</u>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<u style=\\"text-decoration: line-through\\">[bar]</u>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<u style=\\"text-decoration: line-through\\">[bar]</u>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<u style=\\"text-decoration: line-through\\">b[a]r</u>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"text-decoration:line-through\\">b</span>a<span style=\\"text-decoration:line-through\\">r</span>baz" but got "foo<u style=\\"text-decoration:line-through\\">b</u>a<u style=\\"text-decoration:line-through\\">r</u>baz"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<u style=\\"text-decoration: line-through\\">b[a]r</u>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<u style=\\"text-decoration: line-through\\">b[a]r</u>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<u style=\\"text-decoration: line-through\\">b[a]r</u>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<strike>b</strike>a<strike>r</strike>baz" but got "foo<u style=\\"text-decoration:line-through\\">b</u>a<u style=\\"text-decoration:line-through\\">r</u>baz"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<u style=\\"text-decoration: line-through\\">b[a]r</u>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<u style=\\"text-decoration: line-through\\">b[a]r</u>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<s style=\\"text-decoration: overline\\">[bar]</s>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<s style=\\"text-decoration: overline\\">[bar]</s>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<s style=\\"text-decoration: overline\\">[bar]</s>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<s style=\\"text-decoration: overline\\">[bar]</s>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<s style=\\"text-decoration: overline\\">b[a]r</s>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<s style=\\"text-decoration: overline\\">b[a]r</s>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<s style=\\"text-decoration: overline\\">b[a]r</s>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<s style=\\"text-decoration: overline\\">b[a]r</s>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<u style=\\"text-decoration: overline\\">[bar]</u>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"text-decoration:overline line-through\\">bar</span>baz" but got "foo<span style=\\"text-decoration-line:overline line-through\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<u style=\\"text-decoration: overline\\">[bar]</u>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<u style=\\"text-decoration: overline\\">[bar]</u>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<u style=\\"text-decoration: overline\\">[bar]</u>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<strike><u style=\\"text-decoration:overline\\">bar</u></strike>baz" but got "foo<span style=\\"text-decoration-line:overline\\"><strike>bar</strike></span>baz"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<u style=\\"text-decoration: overline\\">[bar]</u>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<u style=\\"text-decoration: overline\\">[bar]</u>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<u style=\\"text-decoration: overline\\">b[a]r</u>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<u style=\\"text-decoration:overline\\">b<span style=\\"text-decoration:line-through\\">a</span>r</u>baz" but got "foo<u style=\\"text-decoration:overline\\">b</u><span style=\\"text-decoration-line:overline line-through\\">a</span><u style=\\"text-decoration:overline\\">r</u>baz"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<u style=\\"text-decoration: overline\\">b[a]r</u>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<u style=\\"text-decoration: overline\\">b[a]r</u>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<u style=\\"text-decoration: overline\\">b[a]r</u>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<u style=\\"text-decoration:overline\\">b<strike>a</strike>r</u>baz" but got "foo<u style=\\"text-decoration:overline\\">b</u><span style=\\"text-decoration-line:overline\\"><strike>a</strike></span><u style=\\"text-decoration:overline\\">r</u>baz"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<u style=\\"text-decoration: overline\\">b[a]r</u>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<u style=\\"text-decoration: overline\\">b[a]r</u>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<p style=\\"text-decoration: line-through\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"text-decoration:line-through\\">foo</span>bar<span style=\\"text-decoration:line-through\\">baz</span></p>" but got "<p style=\\"\\"><span style=\\"text-decoration-line:line-through\\">foo</span>bar<span style=\\"text-decoration-line:line-through\\">baz</span></p>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<p style=\\"text-decoration: line-through\\">foo[bar]baz</p>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<p style=\\"text-decoration: line-through\\">foo[bar]baz</p>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<p style=\\"text-decoration: line-through\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><strike>foo</strike>bar<strike>baz</strike></p>" but got "<p style=\\"\\"><span style=\\"text-decoration-line:line-through\\">foo</span>bar<strike>baz</strike></p>"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<p style=\\"text-decoration: line-through\\">foo[bar]baz</p>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<p style=\\"text-decoration: line-through\\">foo[bar]baz</p>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<p style=\\"text-decoration: overline\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"text-decoration:overline\\">foo<span style=\\"text-decoration:line-through\\">bar</span>baz</p>" but got "<p style=\\"\\"><span style=\\"text-decoration-line:overline\\">foo</span><span style=\\"text-decoration-line:overline line-through\\">bar</span><span style=\\"text-decoration-line:overline\\">baz</span></p>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<p style=\\"text-decoration: overline\\">foo[bar]baz</p>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<p style=\\"text-decoration: overline\\">foo[bar]baz</p>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<p style=\\"text-decoration: overline\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"text-decoration:overline\\">foo<strike>bar</strike>baz</p>" but got "<p style=\\"\\"><span style=\\"text-decoration-line:overline\\">foo<strike>bar</strike>baz</span></p>"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<p style=\\"text-decoration: overline\\">foo[bar]baz</p>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<p style=\\"text-decoration: overline\\">foo[bar]baz</p>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<span class=\\"underline\\">[bar]</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span class=\\"underline\\" style=\\"text-decoration:line-through\\">bar</span>baz" but got "foo<span class=\\"underline\\" style=\\"text-decoration-line:line-through\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<span class=\\"underline\\">[bar]</span>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<span class=\\"underline\\">[bar]</span>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<span class=\\"underline\\">[bar]</span>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<span class=\\"underline\\">[bar]</span>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<span class=\\"underline\\">b[a]r</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span class=\\"underline\\">b<span style=\\"text-decoration:line-through\\">a</span>r</span>baz" but got "foo<span class=\\"underline\\">b<span style=\\"text-decoration-line:line-through\\">a</span>r</span>baz"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<span class=\\"underline\\">b[a]r</span>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<span class=\\"underline\\">b[a]r</span>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<span class=\\"underline\\">b[a]r</span>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<span class=\\"underline\\">b[a]r</span>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "foo<span class=\\"line-through\\">[bar]</span>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "foo<span class=\\"line-through\\">[bar]</span>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "foo<span class=\\"line-through\\">b[a]r</span>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "foo<span class=\\"line-through\\">b[a]r</span>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "foo<span class=\\"underline-and-line-through\\">[bar]</span>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "foo<span class=\\"underline-and-line-through\\">[bar]</span>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "foo<span class=\\"underline-and-line-through\\">b[a]r</span>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "foo<span class=\\"underline-and-line-through\\">b[a]r</span>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "fo[o<s>b]ar</s>baz" queryCommandIndeterm("strikethrough") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""]] "fo[o<s>b]ar</s>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""]] "fo[o<s>b]ar</s>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "foo<s>ba[r</s>b]az" queryCommandIndeterm("strikethrough") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""]] "foo<s>ba[r</s>b]az" queryCommandState("strikethrough") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["strikethrough",""]] "foo<s>ba[r</s>b]az" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "foo<s>ba[r</s>b]az" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""]] "foo<s>ba[r</s>b]az" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "fo[o<s>bar</s>b]az" queryCommandIndeterm("strikethrough") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "fo[o<s>bar</s>b]az" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "fo[o<s>bar</s>b]az" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "fo[o<s>bar</s>b]az" queryCommandIndeterm("strikethrough") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "fo[o<s>bar</s>b]az" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "fo[o<s>bar</s>b]az" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo[<s>b]ar</s>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo[<s>b]ar</s>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo[<s>b]ar</s>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo[<s>b]ar</s>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<s>ba[r</s>]baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<s>ba[r</s>]baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<s>ba[r</s>]baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<s>ba[r</s>]baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo[<s>bar</s>]baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo[<s>bar</s>]baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo[<s>bar</s>]baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo[<s>bar</s>]baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<s>[bar]</s>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
Harness: the test ran to completion.
